fix(bmp180): fix Linux test failures across all platforms
cpp:
- Add #include <cmath> and Linux-compatible delay() shim (usleep) to BMP180.cpp
- Fix _compensate_temp/_compensate_pressure signature: uint16_t/uint32_t → int32_t
  to match .cpp definitions
- Rewrite test to use public API only (protected fields/methods not accessible)

python:
- Remove erroneous SPITransport/SMBusTransport re-export from periph/__init__.py;
  platform-specific transports are imported directly by users

rust:
- Fix calibration parsing: single-byte reads → i16/u16::from_be_bytes (was reading
  only the high byte, causing B2=4 to be read as 0 and fail validity check)
- Change #![no_std] → #![cfg_attr(not(feature = "std"), no_std)] so the std feature
  actually enables std::thread::sleep in delay_ms
- Enable features = ["std"] in bmp180_test Cargo.toml so delays are real
- Fix test_linux.sh path: tests/$TARGET → tests/$CATEGORY (wrong nested path)

+++ cpp/src/chips/pressure/BMP180.h
@@ -76,8 +76,8 @@ protected:
     void     _write_reg(uint8_t reg, uint8_t value);
     uint16_t _read_raw_temp();
     uint32_t _read_raw_pressure();
-    int32_t  _compensate_temp(uint16_t ut);
-    int32_t  _compensate_pressure(uint32_t up);
+    int32_t  _compensate_temp(int32_t ut);
+    int32_t  _compensate_pressure(int32_t up);
     float    _compensate_temp_f(int32_t b5);
     float    _compensate_pressure_f(int32_t up);
 };
